+++ index.html
@@ -43,6 +43,6 @@
     </main>
 
 
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </body>
 </html>
